fix: split login and record
