+++ calypsokit/calydb/record.py
@@ -1,83 +1,20 @@
-import os
-import pickle
 import warnings
 from collections import UserDict
 from time import localtime, strftime
+from datetime import datetime
 
-import dotenv
 import numpy as np
-import pandas as pd
-import pymongo
-from bson import Binary
-from bson.binary import USER_DEFINED_SUBTYPE
-from bson.codec_options import TypeCodec, TypeRegistry, CodecOptions
 
 
-# =========== Registor Numpy Type ==========
-class NumpyCodec(TypeCodec):
-    python_type = np.ndarray
-    bson_type = Binary
-
-    def transform_python(self, value):
-        return Binary(pickle.dumps(value, protocol=2), USER_DEFINED_SUBTYPE)
-
-    def transform_bson(self, value):
-        if value.subtype == USER_DEFINED_SUBTYPE:
-            return pickle.loads(value)
-        return value
-
-
-def fallback_encoder(value):
-    if isinstance(value, np.ndarray):
-        return Binary(pickle.dumps(value, protocol=2), USER_DEFINED_SUBTYPE)
-    return value
-
-
-def get_codec_options():
-    numpy_codec = NumpyCodec()
-    type_registry = TypeRegistry([numpy_codec], fallback_encoder=fallback_encoder)
-    codec_options = CodecOptions(type_registry=type_registry, tz_aware=False)
-    return codec_options
-
-
-def get_collection(name, db):
-    codec_options = get_codec_options()
-    return db.get_collection(name, codec_options=codec_options)
-
-
-def connect_to_db(addr=None, user=None, pwd=None, db=None, col=None, dotenv_path=None):
-    dotenv.load_dotenv(dotenv_path=dotenv_path, override=True)
-    addr = os.environ.get('MONGO_ADDR', None) if addr is None else addr
-    user = os.environ.get('MONGO_USER', None) if user is None else user
-    pwd = os.environ.get('MONGO_PWD', None) if pwd is None else pwd
-    db = os.environ.get('MONGO_DB', None) if db is None else db
-    col = os.environ.get('MONGO_COLLECTION', None) if col is None else col
-    for key in (addr, user, pwd, db, col):
-        if key is None:
-            raise ValueError(f"{key} not configured in .env file")
-
-    client = pymongo.MongoClient(f"mongodb://{user}:{pwd}@{addr}")
-    db = client[db]
-    col = get_collection(col, db)
-
-    all_index = col.index_information()
-    if all_index.get("material_id", False):
-        col.ensure_index([("material_id", 1)], unique=True)
-    else:
-        col.create_index([("material_id", 1)], unique=True)
-
-    return db, col
-
-
-class DocDict(UserDict):
+class BaseRecordDict(UserDict):
     def update_time(self):
-        self.data["last_update"] = strftime("%d/%M/%Y %H:%M:%S %z", localtime())
+        self.data["last_updated_utc"] = datetime.utcnow()
 
     def todb(self):
         return self.data
 
 
-class RawDocDict(DocDict):
+class RecordDict(BaseRecordDict):
     def __init__(self, initdata=None):
         super().__init__(initdata)
         default_data = {
@@ -117,10 +54,15 @@ class RawDocDict(DocDict):
                 "version": "",
                 "ICODE": 0,
             },
-            "incar":                    [],   # list of str，不要替换换行符
-            "potcar":                   [],   # list of str，不要替换换行符
+            "abinitconfig":             [],   # list of str，不要替换换行符
+            "pseudopotential":          [],   # list of str，与elements对应
             "symmetry": {
                 "1e-1": {                       # str, symprec, %.0e
+                    "number":         0,        # int [1, 230]
+                    "symbol":         "",
+                    "crystal_system": "",
+                },
+                "1e-2": {                       # str, symprec, %.0e
                     "number":         0,        # int [1, 230]
                     "symbol":         "",
                     "crystal_system": "",
@@ -131,8 +73,6 @@ class RawDocDict(DocDict):
                 "name": "",
                 "email": "",
             },
-            "mtime":              strftime("%d/%M/%Y %H:%M:%S %z", localtime()),
-            "timefmt":            "%d/%M/%Y %H:%M:%S %z",
             "deprecated":               False,  # bool
             "deprecated_reason":        "",     # str
         }  # fmt: skip
